Split off MMCS code
